Assignments/mpi/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
h11==0.14.0
idna==3.4
mpi4py==3.1.4
passlib==1.7.4
pydantic==1.10.7
PyJWT==2.7.0
PyYAML==6.0
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
unicorn==2.0.1.post1
urllib3==2.0.2
uvicorn==0.22.0
yml==0.0.1
